requirements.txt
Django==3.2.8
djangorestframework==3.14.0
Faker==15.3.4
gunicorn==20.1.0
whitenoise==6.2.0
